Re-added basic ELF loader

+++ kernel/include/klib.h
@@ -13,7 +13,7 @@ char* strncpy(char* dest, const char* src, size_t maxLength);
 void strcpy(char* dest, const char* source);
 size_t strcmp(const char* x, const char* y);
 size_t pow(const size_t number, const size_t power);
-void memset(void* b, int c, int len);
+void memset(void* b, int c, size_t len);
 void memcpy(void *dest, void *src, size_t n);
 
 // To keep intellisense happy
